// src/api/openapi/model/offerElementField.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferField } from './offerField';
import { OfferLibrary } from './offerLibrary';


export interface OfferElementField { 
    defaultValue: string;
    field: OfferField;
    id: number;
    inherits: boolean;
    library?: OfferLibrary | null;
    mandatory: boolean;
}